+++ UnifiedDiagnostics/config.py
@@ -23,7 +23,7 @@ APPEARANCE_MODE: str = "Dark"
 COLOR_THEME: str = "blue"
 
 # Window
-APP_VERSION: str = "1.1.0"
+APP_VERSION: str = "1.2.0"
 WINDOW_TITLE: str = "Master Sentinal"
 WINDOW_GEOMETRY: str = "1100x700"